+++ importers/finmind/exceptions.py
@@ -1,0 +1,14 @@
+class FinMindError(Exception):
+    pass
+
+
+class FinMindAPIError(FinMindError):
+    pass
+
+
+class FinMindRateLimitError(FinMindAPIError):
+    pass
+
+
+class FinMindAuthenticationError(FinMindAPIError):
+    pass